unify model predict config, change dir tree structure, add AutoModelWithLMHead
